feat: Update dark theme colors, enhance HomeScreen layout, and add ChatTextField and QuickPrompts widgets
