+++ lib/main.dart
@@ -1,10 +1,9 @@
-import 'dart:ui';
-import 'package:egypto_ai/config/theme/dark.dart';
-import 'package:egypto_ai/locator.dart';
-import 'package:egypto_ai/utils/helpers/router.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:egypto_ai/utils/helpers/router.dart';
+import 'package:egypto_ai/config/theme/dark.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
+import 'package:egypto_ai/locator.dart';
 import 'package:flutter/material.dart';
 import 'firebase_options.dart';
